Bigger refactors part 2:
- Renamed pluggins to plugins
- Deleted unnecessary files
- Added Pipe decorator to mod.ts
- Modified signature of Controller decorator
