+++ mod.ts
@@ -25,6 +25,7 @@ export { Catch } from "./main-core/decorators/stereotypes/catch/catch.ts";
 export { Override } from "./main-core/decorators/native-components/override.ts";
 export { Cors } from "./mvc-framework/core/decorators/stereotypes/controller/corsMiddlewareDecorator.ts";
 export { Render } from "./mvc-framework/core/decorators/stereotypes/view-engine/viewEngineDecorators.ts";
+export { Pipe } from "./mvc-framework/core/decorators/stereotypes/controller/pipeDecorator.ts";
 export { ViewModel } from "./mvc-framework/core/modules/view-engine/viewModel.ts";
 export { MandarineMVC } from "./mvc-framework/mandarineMVC.ts"
 
@@ -36,8 +37,8 @@ export { Types as SQLTypes } from "./orm-core/sql/types.ts";
 export { ResourceHandler } from "./mvc-framework/core/internal/components/resource-handler-registry/resourceHandler.ts";
 export { MandarineResourceResolver } from "./main-core/mandarine-native/mvc/mandarineResourceResolver.ts";
 
-export { Optional } from "./pluggins/optional.ts";
-export { PromiseRepeater } from "./pluggins/promiseRepeater.ts";
+export { Optional } from "./plugins/optional.ts";
+export { PromiseRepeater } from "./plugins/promiseRepeater.ts";
 
 export { MandarineSessionHandler } from "./main-core/mandarine-native/sessions/mandarineDefaultSessionStore.ts";
