+++ mvc-framework/core/middlewares/resourceHandlerMiddleware.ts
@@ -1,6 +1,5 @@
 // Copyright 2020-2020 The Mandarine.TS Framework authors. All rights reserved. MIT license.
 
-import { Middleware } from "../../../deps.ts";
 import { Mandarine } from "../../../main-core/Mandarine.ns.ts";
 import { ApplicationContext } from "../../../mod.ts";
 import { handleCors } from "./cors/corsMiddleware.ts";
